requirements.txt
alembic
celery
cryptography
fastapi
passlib
PyJWT
PyMySQL
python-dotenv
python-telegram-bot
redis
requests
slowapi
SQLAlchemy
uvicorn